+++ common/src/main/java/com/brokenkeyboard/usefulspyglass/Constants.java
@@ -1,0 +1,11 @@
+package com.brokenkeyboard.usefulspyglass;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class Constants {
+
+	public static final String MOD_ID = "usefulspyglass";
+	public static final String MOD_NAME = "Useful Spyglass";
+	public static final Logger LOG = LoggerFactory.getLogger(MOD_NAME);
+}
