+++ common/src/main/java/com/brokenkeyboard/usefulspyglass/EntityFinder.java
