+++ common/src/main/java/com/brokenkeyboard/usefulspyglass/platform/IPlatformHelper.java
@@ -1,0 +1,8 @@
+package com.brokenkeyboard.usefulspyglass.platform;
+
+import java.util.concurrent.ConcurrentMap;
+
+public interface IPlatformHelper {
+    
+    ConcurrentMap<String, String> getModList();
+}
