+++ common/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+    "pack": {
+        "description": "${mod_name}",
+        "pack_format": 15
+    }
+}
