+++ fabric/src/main/resources/META-INF/services/com.brokenkeyboard.usefulspyglass.platform.IPlatformHelper
@@ -1,0 +1,1 @@
+com.brokenkeyboard.usefulspyglass.platform.FabricPlatformHelper
