+++ fabric/src/main/resources/usefulspyglass.fabric.mixins.json
@@ -1,0 +1,18 @@
+{
+    "required": true,
+    "minVersion": "0.8",
+    "package": "com.brokenkeyboard.usefulspyglass.mixin",
+    "refmap": "${mod_id}.refmap.json",
+    "compatibilityLevel": "JAVA_17",
+    "mixins": [
+        "EnchantmentHelperMixin",
+        "EnchantmentMixin"
+    ],
+    "client": [
+    ],
+    "server": [
+    ],
+    "injectors": {
+      "defaultRequire": 1
+    }
+}
