+++ gradle.properties
@@ -1,15 +1,31 @@
-# Sets default memory used for gradle commands. Can be overridden by user or command line properties.
-# This is required to provide enough memory for the Minecraft decompilation process.
+# Important Notes:
+# fabric.mod.json's modid field cannot be expanded, you must change it manually.
+# Every field you add must be added to the root build.gradle expandProps map.
+
+# Project
+version=0.4.1
+group=com.brokenkeyboard.usefulspyglass
+
+# Common
+minecraft_version=1.20.1
+
+# Neoforge
+forge_version=47.1.64
+forge_loader_version_range=[47.1,)
+forge_version_range=[47.1,)
+minecraft_version_range=[1.20, 1.21)
+
+# Fabric
+fabric_version=0.83.1+1.20.1
+fabric_loader_version=0.14.21
+
+# Mod options
+mod_name=Useful Spyglass
+mod_author=BrokenKeyboard
+mod_id=usefulspyglass
+license=GPLv3
+description=This mod allows players to view information on mobs when viewing them through a spyglass. Also adds spyglass enchantments.
+
+# Gradle
 org.gradle.jvmargs=-Xmx3G
 org.gradle.daemon=false
-
-minecraft_version=1.20.1
-minecraft_version_range=[1.20,1.21)
-neo_version=47.1.64
-neo_version_range=[47.1,)
-
-mod_loader=forge
-mod_id=usefulspyglass
-mod_name=Useful Spyglass
-mod_version=0.4.0
-loader_version_range=[47,)
