--- src/main/resources/META-INF/mods.toml
@@ -1,26 +1,0 @@
-modLoader="javafml"
-loaderVersion="${loader_version_range}"
-license="GPLv3"
-
-[[mods]]
-modId="usefulspyglass"
-version="${file.jarVersion}"
-displayName="Useful Spyglass"
-logoFile="usefulspyglass.png"
-credits="BrokenKeyboard"
-authors="BrokenKeyboard"
-description='''This mod allows players to view information on mobs when viewing them through a spyglass. Also adds spyglass enchantments.'''
-
-[[dependencies.usefulspyglass]]
-    modId="forge"
-    mandatory=true
-    versionRange="${neo_version_range}"
-    ordering="NONE"
-    side="BOTH"
-
-[[dependencies.usefulspyglass]]
-    modId="minecraft"
-    mandatory=true
-    versionRange="${minecraft_version_range}"
-    ordering="NONE"
-    side="BOTH"
