--- src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-    "pack": {
-        "description": "Useful Spyglass resources",
-        "pack_format": 15
-    }
-}
